--- src/utilities/index.js
@@ -1,3 +1,0 @@
-import useStoreState from "./useStoreState";
-
-export default useStoreState;
